Changed notation
gsfc -> obs
calving -> ice_are_extent
