+++ bin/ice_area_extent/GRE_Basins_IMBIE2_v1.3/GRE_Basins_IMBIE2_v1.3.shx
