+++ bin/ice_area_extent/basins/rignot_basins_SW.exp
