+++ bin/ice_area_extent/doi_10_7280_D1WT11__v20190329/Greenland_Basins_PS_v1.4.2.dbf
