+++ cmct/ice_area_extent_modules/legacy/nonnumpy_residual_calculation.py
