+++ cmct/ice_area_extent_modules/legacy/pure_python_residual_calculation.py
