+++ data/ice_area_extent/gsfc_stats.npy
