+++ notebooks/Ice_area_extent/observations_stats.npy
