+++ test/ice_area_extent/ensemble/ENSEMBLE_OUTPUT.npy
